Add variable for file name extension
Initialize this variable with ".eml"
EML files are interpreted by some editors as text files containing e-mail.
On Windows, EML files are opened with Outlook by default.
Some editors (e.g. gVim) will treat EML files specially, e.g. by
breaking lines after 76 (or so) characters.

The separating dot is part of the extension variable, which is not
exactly clean. However, I thought it preferable to having a file end in
a dot (if no extension was specified) or adding program logic to add or
not add a dot or a default extension.

To change back to old behavior, set TEMP_FILE_EXTENSION to ".txt".